backend/requirements.txt
fastapi==0.115.7
kubernetes==31.0.0
uvicorn==0.25.0
python-dotenv==1.0.1
langchain==0.3.15
langchain-chroma==0.2.0
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
openai==1.60.0
psycopg2-binary==2.9.9